requirements.txt
pandas
numpy
matplotlib
seaborn
opencv-python
pillow
plotly
tqdm
scikit-learn
streamlit
